Move route handlers to seperate module
